Some porting progress.
