+++ src/libs/blueprint/conduit_blueprint_mesh_utils.hpp
@@ -23,7 +23,6 @@
 // conduit lib includes
 //-----------------------------------------------------------------------------
 #include "conduit.hpp"
-#include "conduit_blueprint_mesh_topology_metadata.hpp"
 #include "conduit_blueprint_exports.h"
 
 //-----------------------------------------------------------------------------
@@ -198,294 +197,6 @@ public:
 private:
     void init(const ShapeType &shape_type);
 };
-
-//---------------------------------------------------------------------------//
-/**
- The TopologicalMetadata takes an input topology and decomposes each element
- according to a topological cascade to produce connectivity from the input
- topology's level down to the points. In 3D, this gives rise to:
-
-   3D cells, 2D faces, 1D edges, 0D points
- 
- The set of 2D faces is a unique set of faces shared among the input 3D elements
- and mappings are produced that allow us to access for each element which faces
- it "owns" as children. When 2 elements share a face, they will each count it
- as a child in their connectivity. The 2D faces consist of a set of edges
- and these are made unique as well and are available in the mapping.
-
- An example is helpful to understand the concept of the maps.
-
- 4 3D Hex topology (points are labelled):
-
-             |--------------------------- element 2 (3,4,7,6,12,13,16,15)
-             |
-      +-------------+-------------+
-     /|6           /|7           /|8
-    / |           / |           / |
-   /  |          /  |          /  |
-  /   |         /   |         /   |  <--- element 3 (4,5,8,7,13,14,17,16)
- +-------------+-------------+    |
- |15  |        |16  |        |17  |
- |    +- - - - | - -+ - - - -|- - +
- |   /|3       |   /|4       |   /|5
- |  / |        |  / |        |  / |
- | /  |        | /  |        | /  |
- |/   |        |/   |        |/   |
- +-------------+-------------+    |
- |12  |        |13  |        |14  |
- |    +- - - - | - -+ - - - -|- - +
- |   / 0       |   / 1       |   / 2
- |  /          |  /          |  /
- | /           | /           | /  <------ element 1 (1,2,5,4,10,11,14,13)
- |/            |/            |/
- +-------------+-------------+
-  9             10            11 
-        ^              
-        |-------------------------------- element 0  (0,1,4,3,9,10,13,12)
-
-
- When the topological cascade is done, each element is turned into a set of faces
- and these are made unique and elements that abut will share the common face. So,
- elements 0,1 will share the quad face (10,1,4,13). The unique set of faces in all
- elements will form the metadata's dimension 2 connectivity:
-
- face 0: 0, 3, 4, 1
- face 1: 0, 1, 10, 9
- face 2: 1, 4, 13, 10  <----- face 2 is made of these points. face 2 is shared
- face 3: 4, 3, 12, 13         by elements 0, 1.
- face 4: 3, 0, 9, 12
- face 5: 9, 10, 13, 12
- face 6: 1, 4, 5, 2
- face 7: 1, 2, 11
- face 8: 10, 2, 5, 14
- face 9: 11, 5, 4, 13
- face 10: 14, 10, 11, 14
- ...
-
- In the same way, the entire collection of faces can be thought of as a unique
- set of edges with their own numbering.
-
- The entity maps contain the information of how these elements,faces,edges,points
- are connected.
-
- The global entity map for dimension 3 to dimension 2, which we'll write as G(3,2),
- indicates the relationship between entities of dimension 3 to the entities of
- dimension 2. Since the dimension 3 elements are made of the dimension 2 elements
- (much as PH cells are expressed in Blueprint), the G(3,2) map will consist
- of the indices of the faces that make up element 0, followed by the indices of the
- faces that make up element 1, and so on.
-
-           element 0 faces   element 1 faces ...
- G(3,2) = {0, 1, 2, 3, 4, 5}{6, 7, 8, 9, 2, 10}...
-                    |
-                    |
-                    face 3 in unique face list (topo[2] connectivity)
-
- There are other global entity maps G(e,a) too. G(3,3) is the relationship of the elements
- to elements, which is self so any input to the map is also the output. The G(3,1) map
- represents the indices of the 1D edges that make up the 3D element. The G(3,0) map
- represents all of the points contained by the 3D element, values which for most
- cases are contained in the 3D connectivity.
-
- G(e,a) is shorthand for "Global(entity_dim, association_dim)".
-
- When e==a, the relationship is self.
- When e>a, the relationship is that entity e contains/references entity a.
- When a>e, the relationship is entity a is referenced by parents of dimension e.
-
- In the above example for G(3,2), we can see that face 2 is shared by elements 0
- and 1. If we want to know which edges make up face 2, we can consult G(2,1).
-
-           face 0      face 1      face 2      face 3
- G(2,1) = {0, 1, 2, 3}{3, 4, 5, 6}{2, 7, 8, 4}{1, 9, 10, 7}...
-                                   |
-                                   |
-                                   face 2 in unique edge list (topo[1] connectivity)
-
- If we want to know which points make up the edges for face 2, consult G(1,0). This is
- the map of lines to points (also top_dims[1] connectivity).
-
-           edge number
-           |
-           0    1    2    3    4     5     6    7     8
- G(1,0) = {0,3}{3,4}{4,1}{1,0}{1,10}{10,9}{9,0}{4,13}{13,10}...
-                     |         |                |     |
-                     |         |                |     |
-                     ----------------------------------
-                     |
-                     edges that are in face 2. 
-
-
- What does G(e,a) mean?
-
-                               Association (a)
-  ---------------------------------------------------------------------------
-     | 3              | 2               | 1               | 0
-  ---------------------------------------------------------------------------
-E  3 |                | face ids in     | edge ids for    | point ids in
-n    |    self        | element i       | element i       | element i
-t    |                | BBC*            |                 | dim[3] connectivity
-i  --|-----------------------------------------------------------------------
-t  2 | element ids    |                 | edge ids in     | point ids in 
-y    | that contain   |      self       | face i          | face i
-     | face i         |                 | BBC *           | dim[2] connectivity
-e  --|-----------------------------------------------------------------------
-   1 | element ids    | face ids that   |                 | point ids in
-     | that contain   | contain edge i  |      self       | edge i
-     | edge i         |                 |                 | dim[1] connectivity
-   --|-----------------------------------------------------------------------
-   0 | element ids    | face ids that   | edge ids that   |
-     | that contain   | contain point i | contain point i |     self
-     | point i        |                 |                 |
-  ---------------------------------------------------------------------------
-
- *BBC = built by cascade
-
- */
-class CONDUIT_BLUEPRINT_API TopologyMetadata : public detail::NewTopologyMetadata
-{
-public:
-    using ParentClass = detail::NewTopologyMetadata;
-    using IndexType = detail::NewTopologyMetadata::IndexType;
-//    enum IndexType { GLOBAL = 0, LOCAL = 1 };
-
-    //-----------------------------------------------------------------------
-    /**
-     @brief Legacy constructor, which builds all of the topology levels in the shape
-            cascade as well as all associations.
-
-     @param topology     The input topology node.
-     @param coordset     The input coordset associated with the topology.
-     */
-    TopologyMetadata(const conduit::Node &topology, const conduit::Node &coordset) :
-        ParentClass(topology, coordset)
-    {
-    }
-
-    //-----------------------------------------------------------------------
-    /**
-     @brief Constructor for the NewTopologyMetadata class. This constructor
-            lets the caller be more selective about which topology levels and 
-            associations are created, possibly saving time.
-
-     @param topology     The input topology node.
-     @param coordset     The input coordset associated with the topology.
-     @param lowest_dim   The lowest level of shape cascade that we're interested in.
-                         If we only want faces, for example, then we don't need to
-                         refine to edges or points.
-     @param desired      A vector of (entity_dim,assoc_dim) pairs that indicate
-                         the associations that will be requested by the client.
-     */
-    TopologyMetadata(const conduit::Node &topology,
-                        const conduit::Node &coordset,
-                        size_t lowest_dim,
-                        const std::vector<std::pair<size_t,size_t> > &desired_maps) :
-        ParentClass(topology, coordset, lowest_dim, desired_maps)
-    {
-    }
-
-    //-----------------------------------------------------------------------
-    /**
-     @brief Get the highest shape dimension.
-     @return The highest shape dimension.
-     */
-    int dimension() const { return ParentClass::dimension(); }
-
-    //-----------------------------------------------------------------------
-    /**
-     @brief Get the topologies array.
-     @return The topologies array.
-     */
-    const conduit::Node *get_topologies() const
-    {
-        return ParentClass::get_topologies();
-    }
-
-    //-----------------------------------------------------------------------
-    /**
-     @brief Get the legnths arrays for the topologies. Any topologies that
-            were not produced will have length 0.
-     @return The topology lengths array.
-     */
-    const index_t *get_topology_lengths() const
-    {
-        return ParentClass::get_topology_lengths();
-    }
-
-    //-----------------------------------------------------------------------
-    /**
-     @brief Get a global association for a particular entity_id in the entity_dim
-            dimension. This lets us ask a question like: which face ids are in
-            element 5? Or, which faces contain edge 12?
-
-     @param entity_id The entity id whose information we want to obtain. This
-                      value must be valid for the entity dimension in question.
-     @param entity_dim The dimension of the entity whose information we want.
-     @param assoc_dim  The dimension of the association we want.
-
-     Example: Which faces are in element 5?
-              get_global_association(5,  // element 5
-                                     3,  // elements are 3 dimensional
-                                     2); // faces are 2 dimensional
-
-     Example 2: Which faces contain edge 2?
-              get_global_association(2,  // edge 2
-                                     1,  // edges are 1 dimensional
-                                     2); // faces are 2 dimensional
-
-     @note The data are returned as a vector_view, which is an object that acts
-           like a vector but does not have to copy bulk data. The data it points
-           to for a given return are in the associations.
-     */
-    conduit::vector_view<index_t>
-    get_global_association(index_t entity_id, index_t entity_dim, index_t assoc_dim) const
-    {
-        return ParentClass::get_global_association(entity_id, entity_dim, assoc_dim);
-    }
-
-    conduit::range_vector<index_t>
-    get_local_association(index_t entity_id, index_t entity_dim, index_t assoc_dim) const
-    {
-        return ParentClass::get_local_association(entity_id, entity_dim, assoc_dim);
-    }
-
-    //-----------------------------------------------------------------------
-    /**
-     @brief Get whether the association e,a was requested when the object
-            was initialized. If so, the association will exist. Otherwise,
-            the association does not exist.
-
-     @param entity_dim The starting dimension.
-     @param assoc_dim  The destination dimension.
-
-     @return True if the association exists; False otherwise.
-     */
-    bool association_requested(index_t entity_dim, index_t assoc_dim) const
-    {
-        return ParentClass::association_requested(entity_dim, assoc_dim);
-    }
-
-    //-----------------------------------------------------------------------
-    /**
-      @brief Return all of the data for an association by COPYING it into the 
-             provided Conduit node.
-
-      @param type Whether we want GLOBAL or LOCAL data.
-      @param src_dim The source dimension of the desired association.
-      @param dst_dim The destination dimension of the desired association.
-      @param[out] map_node The Conduit node that will contain the copied association data.
-
-      @note This method guarantees that all bulk arrays for values, sizes, and
-            offsets will be index_t.
-
-      @return True if the map exists(was requested); False otherwise.
-     */
-    bool get_dim_map(IndexType type, index_t src_dim, index_t dst_dim, Node &map_node) const
-    {
-        return ParentClass::get_dim_map(type, src_sim, dst_dim, map_node);
-    }
-};
-
 
 //-----------------------------------------------------------------------------
 /// blueprint mesh utility functions
